refactor: move builder.py
